Refactor DepartmentUserRecord create and update methods. Refactor unit tests to reduce unit test monolith.
